tests - bring back file editor input factory test
